Ensure library files updated after include all files option changed
On switching from including all files to choosing a specific set
of files in the Add Client-Side Library dialog the list view was
not being updated. When jquery is selected, and choosing a specific
set of files is enabled, only the min.js file is checked. When
include all files is selected then all files are checked.