Ensure Library Manager logger output is available on opening pad
Logger output would not be recorded until the Library Manager
Output pad window was opened the first time. Now the LogView is
created if it is not already created so logger output is available
on opening the pad window the first time.

Also ensure the LogView is cleared when the last item in the
solution is closed.