Fix unhandled exception in Add Client-Side Library dialog
If a library provider could not find a library an exception was
thrown by the task and it was not used which caused an unhandled
exception to be logged by the IDE.

An unhandled exception has occurred. Terminating Visual Studio? False
Microsoft.Web.LibraryManager.Contracts.InvalidLibraryException: The "j" library could not be resolved by the "cdnjs" provider
  at Microsoft.Web.LibraryManager.Providers.Cdnjs.CdnjsCatalog.GetLibraryAsync (System.String libraryName, System.String version, System.Threading.CancellationToken cancellationToken) [0x00084] in LibraryManager/src/LibraryManager/Providers/Cdnjs/CdnjsCatalog.cs:146