Fix completion window not re-opening after escape key pressed
Switched to using the Xwt.TextEntry instead of the SearchTextEntry.
This seems to fix some of the odd behaviour with the completion,
such as:

Completion window not being displayed when typing after it has been
closed by pressing the escape key.

Enter key not adding the selected item to the text box.